aura/requirements.txt
# Aura - Autonomous LLM Coding Assistant Requirements
# Core Dependencies

# Async and networking
httpx>=0.25.0
pyzmq>=25.0.0

# Data processing and ML
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
networkx>=3.0.0

# File watching and system
watchdog>=3.0.0
psutil>=5.9.0

# AST and code analysis
typing_extensions>=4.5.0

# Logging and configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0